fix: Field style and remove old file
