+++ src/Features/Field/Field.jsx
@@ -1,103 +1,100 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 import Image from 'gatsby-image';
-import Container from '../../Utils/Container';
+import {
+  Card,
+  CardActionArea,
+  CardContent,
+  Typography,
+  Grid,
+} from '@material-ui/core';
+import { renderListSafely } from '../../Hocs';
+import useStyles from './useStyles';
 
-const List = styled.ul`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-  list-style: none;
-  margin: 2rem 0;
-  padding: 0;
-`;
+const Field = ({ lists, title }) => {
+  const {
+    card,
+    cardTitle,
+    subtitle,
+    title: pageTitle,
+    body1,
+    ul,
+  } = useStyles();
 
-const ListItem = styled.li`
-  cursor: pointer;
-  flex: 25%;
-  margin: 1rem 0;
-  max-width: 25%;
-  min-width: 325px;
-
-  img {
-    transition: filter 500ms ease-in;
-  }
-
-  &:hover {
-    text-decoration: underline;
-
-    img {
-      filter: grayscale(1);
-    }
-  }
-`;
-
-const ListItemText = styled.div`
-  flex: 1;
-  padding: 1rem;
-
-  h4 {
-    font-size: 1.443rem;
-  }
-
-  p {
-    font-size: 1rem;
-  }
-
-  p,
-  h4 {
-    margin: 0;
-  }
-`;
-
-const Title = styled.h3`
-  font-size: calc(
-    29.479px + (29.479 - 23.086) *
-      ((100vw - 300px) / (1600 - 300))
+  return (
+    <div>
+      {title && (
+        <Typography variant="h3" className={pageTitle}>
+          {title}
+        </Typography>
+      )}
+      <Grid
+        container
+        spacing={6}
+        component="ul"
+        className={ul}
+      >
+        {lists.map((img, i) => (
+          <Grid
+            item
+            component="li"
+            key={i}
+            xs={12}
+            sm={6}
+            md={4}
+            lg={3}
+          >
+            <Card
+              onClick={img.onClick}
+              onKeyPress={img.onClick}
+              className={card}
+              tabIndex={0}
+            >
+              <Image
+                fluid={img.fluid}
+                style={{ width: '100%', height: '200px' }}
+              />
+              <CardContent>
+                {img.subtitle ? (
+                  <span className={subtitle}>
+                    {img.subtitle}
+                  </span>
+                ) : null}
+                <Typography
+                  variant="h4"
+                  className={cardTitle}
+                >
+                  {img.title}
+                </Typography>
+                <Typography
+                  variant="body1"
+                  className={body1}
+                >
+                  {img.description}
+                </Typography>
+              </CardContent>
+            </Card>
+          </Grid>
+        ))}
+      </Grid>
+    </div>
   );
-  margin: 0;
-  text-align: center;
-`;
-
-const Field = ({ data, title }) => (
-  <Container size="large">
-    {title && <Title>{title}</Title>}
-    <List>
-      {data.map((img, i) => (
-        <ListItem
-          tabIndex="0"
-          onClick={img.onClick}
-          onKeyPress={img.onClick}
-          key={i}
-        >
-          <Image {...img} />
-          <ListItemText id={img.title}>
-            <h4>{img.title}</h4>
-            <p>{img.description}</p>
-          </ListItemText>
-        </ListItem>
-      ))}
-    </List>
-  </Container>
-);
+};
 
 Field.propTypes = {
-  data: PropTypes.arrayOf(
+  lists: PropTypes.arrayOf(
     PropTypes.shape({
       title: PropTypes.string,
       description: PropTypes.string,
       src: PropTypes.string,
       onClick: PropTypes.func,
     }),
-  ),
+  ).isRequired,
   title: '',
 };
 
 Field.defaultProps = {
-  data: [],
   title: PropTypes.string,
 };
 
-export default Field;
+export default renderListSafely(Field);
