+++ src/Features/Field/Field.stories.mdx
@@ -4,7 +4,7 @@ import {
   Preview,
   ArgsTable,
 } from '@storybook/addon-docs/blocks';
-import Field from './_Field';
+import Field from './Field';
 import blogList from '../../../cypress/fixtures/blogList';
 
 <Meta title="Features/Field" component={Field} />
