+++ src/Lists/Downloads/useStyles.js
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  title: {
+    fontWeight: 'bold',
+  },
+}));
+
+export default useStyles;
